Sharing base options between ReSharper versions.
